[qfcc] Make __VA_ARGS__ an actual macro parameter
It seems I hadn't understood the C spec very well when I implemented
__VA_ARGS__ as reading it again now was (after a bit of extra thought)
helpful in realizing that __VA_ARGS__ is simply a proper name for the
... "parameter". Fixes __VA_ARGS__ not expanding in the arguments to
another macro.

I think the __VA_ARGS__ magic macro is no longer needed, but I need to
test it properly before deleting it outright.