[sw] Offset stverts correctly for sub-models
The code needs to be cleaned up so there's less hard-coding, but this
gets Mr Fixit working properly (without animations). That is, his head
is rendered correctly and clipping no longer causes segfaults. As I
suspected, it was the mangled UV coordinates (which were due to the
incorrectly offset stverts) causing the segfaults.