[qfcc] Add support for the DebugPrintf spir-v instruction
Or really, include its grammar json and support ... on intrinsics (and
of course in spir-v) as well as string constants in spir-v.

Oh... how I have missed the humble printf
```
validation layer: info: vkQueueSubmit(): pSubmits[0] DebugPrintf:
MB:26391 VI:0
```

As I had hoped, the problem is in my MatrixBase, not ViewIndex.