[sw] Implement bone deformation
However, it's not working too well. I'm not sure if the data is bad (in
the blend palette), the modified indices are bad, or there's a bug in
the application of the matrices. Mr Fixit is a little big for debugging
this kind of thing. Unless the model is forced to use triangle clipping,
the renderer will crash.