[ui] Improve control over background fills
I decided I didn't want to have to always use flexible space to fill
empty regions of layouts. This can make for more overdraw, but I suspect
that'll be less of a problem than I expect, and it does reduce the
number of views needed to do certain widgets (and thus fewer entities
and components).