[ui] Move imui_state_t into the header
The main reason I had it private was the next/prev pointers which I
didn't want exposed, but adding a little wrapper around the struct did
the trick. If I move to using IDs for the strings and other window
pointers, I should be able to make the header shareable with ruamoko.