[ui] Add a scroller to make scrollable contents
It's based on the code for passages but more general. The size and
position can be tweaked to allow for optimal population of the contents
(ie, there's no need to render 1000 items when only 10 are visible). I'm
not super happy with how I get the size of the lines in filewindow, but
it works for now.