[model] Force uniform scale on non-leaf joints
The uniform scale is the average of the three components. Cube root of
the product might be better, but computing that...

Fixes the horrible displacements in my new test model (I think the
mixamo animation has stuff for emulating muscles which got translated
into what seems to be volume-preserving non-uniform scale on some bones,
and those bones got exported as non-leaf bones).