[model] Implement animation (de)registration
Both clips and armatures are registered (for mesh models). The current
implementation is a very iffy, but it at least doesn't crash.

I've also created a template file for using PR_RESMAP because I got
tired of doing all that boilerplate. It should be easy enough to
retrofit to other RESMAP users.