[model] Throw in some Tracy profiling zones
And speed up qfa_update_clip a little. It needs some more work because
the fbsearch is surprisingly expensive (compared to just stepping
through the keyframes when moving forward or backward), but it should
still be good for seeking (and is currently fast enough), so I'll leave
that for now.