[model] Optimize the animation blending
That was an interesting one. My initial implementation of the blend loop
used a naive cast of the data and wound up 30x slower than the scalar
(VectorMultAdd etc) version. After examining the assembly for both
versions, and failing to make a dent in the performance of the simd
version, I realized it might be the name and parent fields causing grief
due to denormals (and nans for parent on a root bone). Sure enough,
fixing that resulted in a 1.5x speedup of the simd blend code (despite
the extra step of zeroing the name and parent fields in the simd
registers).

scalar     : ~ 150ns
naive simd : ~ 4.5us
masked simd: ~ 90ns