[vulkan] Start implementing Guerilla's UIPainter
https://www.guerrilla-games.com/read/uipainter-tile-based-ui-rendering-in-one-draw-call
Or more realistically, my interpretation of it. This is where I found
the need to implement switch for spir-v, since I don't want to use an
if/else block when I get a bunch of commands.

painter_draw is currently in vulkan_draw, but I don't think I want it to
stay there. However, it's enough to get qwaq-ed running without
crashing.