[ui] Support utf-32 text, as well as attributes
For now, only utf-32 text *with* attributes is supported by IMUI itself,
but the lower-level code (text shaping for simple strings) takes care of
utf-32 (without knowing anything about attributes beyond what was
already there).

The attributes support a 64 color palette (not yet implemented) as well
as a single background color (for highlighting). There are problems with
the highlighting in that the views getting the background fill have the
wrong sizes. The wrong sizes also affect text layout as the x size is
often too small as it uses the wrong metric.