[ui] Always prefer advance for text size
Unless it's 0, of course. It turns out that different freetype builds
(eg, x86 vs arm) or versions (I haven't checked) produce slightly
different glyph rects for the same font. This resulted in the width of X
appearing 1 pixel wider than it actually is (or really, 3/4 of a pixel
(though I seem to remember it being even closer to 13 on my desktop)).