[x11] Support XIM for text input
When "focused input" is enabled, XIM is used to process the keyboard
input to allow multi-lingual text. Plain ASCII text is passed as before,
but anything outside of ASCII is now passed via the new utf-8 support in
key events. It's possibly a bit iffy as in_x11 maintains ownership of
the dstring, but the idea is the event receiver would either process the
text immediately or copy it into storage. It saves worrying about
freeing the memory if the event is ignored.