[qwaq] Add insertMsgBuf to EditBuffer
It's assumed that the buffer is a valid, non-empty, utf-8 string that is
nul-terminated with the nul included in the buffer's size. The number of
lines in the buffer is returned so that tracking line count updates is
easier (and there's no nice way to count lines in the msgbuf in
ruamoko).