[qfcc] Ensure emitted obj structs have unique types
That is, those that need unique types, of course. Many of the structs
have variable-length arrays and thus are very specific. However, the
mechanism for generating tags (which was meant to ensure unique types)
doesn't work when those struct are essentially all on the same line (the
last line of the file). I'm pretty sure it was the switch to hashing the
type encoding that resulted in the breakage, which I noticed while
perusing global/type info pondering using such for serialization via
plists/json.