[qfcc] Check source line is in function being emitted
This fixes the broken source lines. The problem is caused by expressions
in macros recording the actual source location of the macro (which is a
good thing), but this breaks the naive line number handling. This fix
causes line number info from macros getting lost (until I can come up
with a solution), but that's better than losing line number info for
huge chunks of code.