[qfcc] Dereference the IMP type for untyped messages
build_function_call expects a bare function type, not a function pointer
(call_expr takes care of the dereferencing when necessary, so this is
equivalent behavior). Fixes the "too many parameters" error (which is
really UB) when sending untyped messages (ie, those to which the class
might not respond thus the types can't be determined reliably).