[qfcc] Don't treat address as killing a node
As the FIXME suggested, taking the address of a variable being
conditional was not the best. Simply removing that fixed the broken
bones display in qwaq-ed (there's now a test for that), but broke other
things, so something was necessary. I decided to make an explicit kill
which works well for .argN, but the attached identifier idea might still
be needed.