[qfcc] Create a new leaf node for address operands
I realized that taking the address of a variable should never kill its
value. This doesn't fix ptr-array in ptrincinout2.r (that's in the
expression tree), but it improves generated code slightly. However, I
think checks need to be made for `lea (ptr + offset)` which shouldn't
create a new leaf node, but that's another "later" as things seem to
work nicely.