[qfcc] Record ud-chains for ambiguous defines
That is, function call reference parameters and dereferenced pointer
assignments. The idea is to fix the refinit test, but it doesn't work
yet (fortunately, everything else still does). I think it's because the
reaching definitions data needs to be recalculated.

However, the ud and du chains seem to be correct.