[qfcc] Use ud-chains for uninitialized variable detection
This greatly simplifies the code and gets refinit working. While
reaching defs do still need to be updated, this does the job for now. I
suspect I'll need to record whether the defs are ambiguous (via
pointers) or unambiguous (direct assignment) but all tests pass and the
generated code does the right thing for now.