[qfcc] Split the dot product in vec * tvec
This makes for better code in general as it allows for more
opportunities where the result cancels out. Fixes the incorrect type for
the result of `(drot • p) * p` where `drot` is a bivector and `p` is a
trivector (the result included a quadvector because of the failure to
cancel the terms).