[gl] Work around the gcc-15 warning
Since everything seems to be working, it seems to be a false-positive
(unless it's something to do with vec[3] not being used and thus the
optimizer trimming the load and then simd using that sub-register and
thus triggering the warning, which I guess makes the warning legit if
garbage winds up in that register causing nans, infs or denormals to be
processed causing a slowdown). Anyway, strange issue but at least I've
got a workaround and everything else seems good (all tests pass, nq and
qwaq-ed work).