[vulkan] Rework end of frame steps and tasks
The command submission is split up so the start of the main render
doesn't have to wait until the swapchain image has been acquired. Some
things are a little messy, but things are working and there might be a
slight improvement in performance, not sure because the actual time
difference is pretty small (but one of the tests did show less weird
gaps on the GPU).