[model] Improve end-of-clip behavior
If the animation time is in the past, or the clip has never played,
start it now, otherwise try to keep the animation smooth by maintaining
phase. This might be a problem if the animation time is too far in the
future, but that can be fixed later if need be.