[model] Implement root motion extraction
A new "model" is created that contains the animation data for only the
root bone of the original model.

The original model is modified such that its root transform is identity
and is no longer animated.

I'm not yet decided on the interface, and the extracted root motion
data hasn't been tested yet, but the modified source model seems to be
working nicely.