[vulkan] Always call lighting_rewrite_ids
This isn't really the right fix as lighting_rewrite_ids does too much by
both rewriting the ids and building the stage queue. It's the fact that
lighting_rewrite_ids wasn't getting called (and thus the stage queue not
getting built) that resulted in no shadows when there were no point or
spot lights in the PVS (or in the case of qwaqed, the scene itself).