[qfcc] Reorganize enum grammar a little
While trying to implement attributes for enums (with the plan of adding
a namespace attribute), I discovered that I got the grammar wrong in
that the backing type comes before the tag instead of after (for
compatibility with C). This just puts all the rules together and uses
named attribtes in enum_specifier.