[input] Support unconnected device axes in keyhelp
The problem was only connected devices had a separate axis info that
wasn't updated directly by the driver. The solution was simply to treat
all devices as unconnected and create a temporary axis info block for
them, which of course recorded the axis values at the time the keyhelp
command was issued.