[vulkan] Use the caster and receiver bounds for CSM
This is a major step towards implementing the algrithm in
`Parallel-Split Shadow Maps`. I think all that's left for that is
computing the split planes, which is interesting with infinite far plane
perspective. Again, I had a lot of trouble getting the matrices right,
due mostly to the z-axis (I like my reverse depth buffers, even for
orthographic views).