[vulkan] Rework the split plane calculation
Now the blended logarithmic splits from the GPU Gems article are used.
However, they've shown a few problems: four cascades is simply not
enough (resolution issues), and there's a problem with the bounds for
the first split: objects behind the camera are getting clipped when they
shouldn't be, resulting in missing shadows.

Regarding the number of cascades: the calculation of the splits can
handle any number, but the shader code is locked at 4, but it shouldn't
be too difficult to fix: I have prototype code elsewhere that does a
nice job of handling any number of cascades.