[qfcc] Make a separate language for qc
The idea was to make @overload optional in Ruamoko, but that proved a
little problematic and I didn't feel like sorting out all the issues
with it just now. However, I figured that keeping the separate language
settings would be handy (save me from having to do it again later).