[vulkan] Convert the compose fragment shader to ruamoko
It itself, it didn't really gain much, but qfcc is a little better
(ruamoko can do a little more on the gpu), and the beginnings of
"standard" definitions for builtins/intrinsics are in place (though
mostly glsl-compatible, for good or ill). Also, the compose and lighting
input attachment layouts are the same other than compose adds the light
buffer to the set (it looks like that maxDescriptorSetInputAttachments
might not be a problem despite the minimum limit being only 4: very rare
device count, even limiting to android, less than 8%).