[qfcc] Correctly handle aliasing alias expressions
Alias expressions are meant to never be more than one deep.
new_offset_alias_expr tried to prevent this but mistakenly checked for
an offset before dealiasing the source expression, thus accidentally
creating a double-alias. This fixes both a triple alias I got when
messing with duals, and non-cancellation in a sandwich product when I
tried to fix that in the algebra code.