[qfcc] Use a simple id for temp operands
I got tired of dealing with 64-bit pointers as the id for temps. Simply
assigning a monotonically incrementing number makes a world of
difference. It uses the same counter as final code generation, but the
counter is reset between generating the statements and emitting the
code, so temp names stay small.