[qfcc] Defer more assignments to algebra components
Really, just single group objects. I'd forgotten about these when doing
the similar patch earlier. Unfortunately, the pga2d test fails due to
alignment issues causing a segfault (a dvec3 temp is used as the
destination to a summed extend pair which means a dvec2 is written, but
dvec2 requires alignment of 4 but dvec3 requires only 2).