[qfcc] Add a failing test for multi-alias expressions
This is triggered by geometric algebra, but I imagine it'll come up with
other complex code involving aliases. The test itself is, technically, a
bit of UB in that it relies on filling the stack with NaN, which then
gets propagated into the result, but it does the trick nicely.