[qfcc] Attach all exact-match alias labels to a node
Exact-match meaning same offset and size (but of course, different type,
otherwise they wouldn't be aliases). This makes it so assigning to one
alias makes the value available to all other aliases that are just
bit-casts of the first alias, thus searching for a node for the later
aliases finds the correct node instead of simply creating a new leaf
node. Fixes algseq3 (and thus the lookat code in qwaq-ed).