[qwaq] Match yaw compensation axis to actual yaw
This moves the sign-flip in the yaw rotor from about -120 to 180 (ie,
anti-parallel). At least that issue makes sense now, especially the fact
that the transition around +120 degrees did not show any signs of
flipping signs. Everything was due to the 180 degree compensation
rotation being applied in the same direction regardless of which way was
the shorter path for the actual rotation.