[qfcc] Don't remove labels from killed nodes
It turns out the normjoin issue was due to the node with the multiple
aliases getting killed by a pointer dereference assignment (which itself
is a problem as the pointer can't kill that node, but that's an
optimization miss, not an actual error) and then the labels for `dir:0`
getting removed so it was never assigned. This change fixes the problem
but now there's another no-op assignment to sort out. It might be time
to think about doing a simple peephole optimizer.