[qfcc] Handle aliasing of unsized arrays
External arrays can be unsized at the time of declaration, which results
in their size being 0 when doing various checks. Skipping the size check
when the size is not yet known avoids an ICE (since those errors are
really for me doing things in qfcc, not errors in the code being
compiled).