[qfcc] Fix automatic group names
It turns out the names were mismatched from the actual groups for any of
the non-specialized algebras, so now the group name is stored as part of
the group for the specialized algebras and grade-relevant names are used
for up to 4d. Beyond that, just grade_G, (G=grade), though I might use
that (as an alias?) for all grades.

Fixes some confusion I found when checking algebra types for R(1,0,0)
and R(2,0,0).