[qfcc] Rework GA expressions to avoid offset aliases
Instead, field expressions and swizzles are used. This had the nice side
effect of making the dot graphs easier to read, too, but the main goal
was spir-v support. There are just a couple of issues preventing
generating spir-v GA code, but there are some validation issues anyway.
However, this was the main hurdle.