[qfcc] Partially optimize dot expressions
This takes care of scalar triple products (a.bxc, or a^b^c), either
cancelling or collecting, as appropriate. The qvec fixme in algseq2
isn't fixed by it because it's due to vector triple products and
corresponding dot products, but it does take care of sandwiching a
bivector with another bivector (180 degree rotation of a line around
another line).