[qfcc] Optimize vector triple products
The following substitutions are made:
    (a×b)×c - a•c b -> -b•c a
    (a×b)×c + b•c a ->  a•c b
    a×(b×c) - a•c b -> -a•b c
    a×(b×c) + a•b c ->  a•c b

This results in faster code when the simple expression isn't canceled
out (2 ops vs 4), but it is very likely the simple expression will be
canceled as the larger expressions tend to come from triple geometric
products. This takes care of most of the (lack of) cancellation issues.