[ruamoko] Check listener is owned before touching its data
I had never mixed use of input buttons and axes between engine core and
progs before, but using imt switchers (which qwaq now does) registers
its own listener with the progs-created button. This resulted in the
ruamoko code trampling on the switcher's data due to assuming it owned
all the listener callbacks. This is the bug that was hiding behind the
deleted input bug.