[qfcc] Use vector construction to merge summed extends
And get rid of the non-multivector exception (merging summed extends is
now done in the multivector assignment code). This is a huge step
towards making geometric algebra SSA compatible, which will make
supporting it in spir-v much easier.

Unfortunately, this breaks qwaq's player camera again, but it turns out
that's due to a weird bug with the arguments to a function the result of
which is unused (ie, removing that basically useless line fixes the
camera again).