[qfcc] Add a failing test for tangled arguments
This is the breakage in qwaq's player camera. It turns out (not very
surprisingly) that it doesn't mater whether the result of the call is
used, but it seems that things get royally messed up for both the log
and the calculation of `pp`.

This lets me remove the offending code from qwaq :)