[qfcc] Rework algebra assignments to be SSA
At least, until the ruamoko and v6 targets get hold of the compound
initializer.

For some reason, this resulted in the ULP errors in brokenplane and
brokenkill changing by 1 (for the worse, unfortunately, but there is a
lot of math going on in there) and I don't feel like chasing that down
at this stage.