[qfcc] Build algebra constructs correctly
Nicely, I'd found a way to share the element chain building between
spirv and ruamoko. All the tests pass for ruamoko and nq and qwaq seem
to be ok with the shaders so it seems there wasn't anything broken in
the process. Further testing will tell.

With this, my little test "shader" passes validation and the code looks
acceptable (it's a little wasteful but does look correct).