[qfcc] Handle vector expressions in geometric algebra
Currently, they need to be constructed "correctly" for optimal code
(though the rest of the code generation doesn't do too badly with
sub-optimal cases). This makes constructing points and planes from vec3
and a suitable scalar (especially 0 or 1) produce very nice code due to
the constant folding getting to have an effect.