[qfcc] Allow the dot product to be used on scalars
However, only if both operands are scalar, and a warning is given for
non-implicit integer types. This was a bit of a surprise, but since GA
defines the dot product on scalars to simply be scalar multiplication
(and the GA code does so in multivectors anyway), and forcing the dot
product between the dual of two pseudoscalars to be treated as a special
case at the source-code level felt wrong. That is, `⋆M.qvec • ⋆M.qvec`
(in 3d PGA) should be valid and not require `*` instead.