[qfcc] Correct offset calculation for main def
The offset passed to def_overlap is relative to the main def, so
d_offset needs to be 0 when d is the main def. This fixes the
alias_vec_param test, but breaks brokenplane and brokenkill. While
brokenkill's fix was alias related, brokenplane was not, so that's a new
issue for that one, though I suspect they're both closely related.