[qfcc] Don't attach labels across call-barriers
This is the correct way to use call-barriers. My previous attempt caused
problems by a barred variable causing the children of its destination
node to be barred. I expect there might be trouble in the future should
a common sub-expression assigned to an arg cross a call, but there seems
to be a suitable check in place for that already (attaching labels
refuses to create loops, and if that happens for a non-assignment op, an
internal error will be generated). All tests pass again, and qwaqed
seems to work properly (just the broken scrolling caused by
signed/unsigned), though there are a few things I've noticed could be
improved (and maybe a lurking bug).