[qfcc] Improve call-barriers a little
Now unrelated nodes are no longer killed by calls, which takes care of
one of the potential bugs I saw. However, not that surprisingly, there
was a case where attaching a label failed, but fixing it was as simple
as I had hoped it would be, particularly since it was a leaf node (so
just create a new leaf node). There's still the problem with expression
nodes, but I'll worry about that when the time comes. At least I now
*know* that the failure will be caught.