[qfcc] Add a failing test for live aliases
I suspect this has been in my "very hard" basket for a while, but I was
finally able to get a small test that reproduces the problem. The
difficulty is that (for whatever reason) the flow graph is split on two
calls (which in itself is a bug, but not relevant to this since an `if`
block in between the calls would have a similar effect), the write to
len.y in the second block kills the whole assignment to len in the first
(not counting the function prolog block), then the use of len after the
write to len.y does not use the assignment to len in the first block
despite len.x still being valid. Part of the problem is there is no
flow-var for len.x (because it's never used) thus the alias code can't
do anything about it.