[vulkan] Use the right packet for waiting on fences
Most of the changes are various checks (they helped a lot when looking
for the bug) and some cleanup, but the loop for waiting on packets to
free up space in the buffer was using the packet for which space was
being allocated to access the fence and thus corrupted packets and the
open space in the buffer (resulting in an out-of-sequence error). This
was tricky to find because the problem happened with just a 2x2 texture
and thus a 64-byte stating buffer (I really need to clean those up)
which meant that the transfer was usually really fast and it would take
some tricky timing to cause a wait. It was always my night-heron model
that triggered it because it has no textures available but wants 12,
thus the tiny staging buffer getting hammered.  Still, it was a good
test of the system.