[model] Reject bad skins sizes on alias models
However, the criteria is just that the resulting byte size of the skin
is not a multiple of 4. That's sufficient to keep the result of the
model data properly aligned. I'm not sure if the width actually needs to
be a multiple of 4 or not, but this will do until I find out.