[vulkan] Remove all secondary staging buffers
Many of them were a lazy hack to force immediate flushing, and the
others... I really don't know why I thought it was a good idea.
Interestingly this has revealed what looks to be a lot of memory
corruption. This may be the real cause of my grief in startup issues.