[vulkan] Handle interleaved packet acquisitions
I think this is why I had multiple staging buffers: it was just easier
(at the time) than figuring out how to check, even though that proved to
be quite easy this time (though I guess I had more motivation to solve
this particular issue). This fixes the corruption of the scrap data, but
not the device lost error. In retrospect, that's not all that surprising
as it was only 4 subpics in the scrap that were corrupted, and no other
staged data seemed to have been corrupted (going through the first
frame's command buffers in in renderdoc).