[vulkan] Redesign staging's buffer management
The old scheme was nice and simple, but it could not cope with the heavy
use induced by every subsystem using the one staging buffer (it would
get its knickers in a twist depending on the acquire/release patterns).
This should fix the corruption (I still need to write decent tests), but
it doesn't fix the DL error (I didn't have all that much hope for it),
though it's still possible there's some corruption going on, but I
suspect it's more likely something wrong with my sync primitives.