[qfcc] Use nil for error initializers
I really don't like that semantic errors in initializers result in the
declaration being abandoned as that results in rather bogus follow-on
errors. Declarations using `auto` get a little weird (void, which itself
should be an error, so follow-on errors will be impossible to avoid
entirely, but at least undeclared errors can be).