[qfcc] Catch variables declared void
Since `auto` declarations using error initializers results in them being
`void`, and declaring variables as `void` are both errors, this
effectively kills two birds with one stone: the latter is caught as it
should be, and the former is less likely to result in weird follow-on
errors. Of course, if the intended type of the `auto` declaration is
incompatible with the default type (which is used as the fallback for
`void` declarations), then there will be further semantic errors, but
this should still reduce the noise in general.