[qwaq] Clean up the attachment specs
The only difference between deferred and deferred_cube attachments is
flat/vs cube for the views and the image references for the format. I
found myself editing the wrong set for ops and clear values, so
combining those and using inheritance makes sense. Not done for the main
engine yet as this is all a part of tracking down the true cause of the
device lost error and nq etc at least work when not starting in fisheye.