[vulkan] Add a comment about the device lost error
I've spent a week chasing this problem and it's hard to say whether I've
made any progress on it (though QF has improved in general due to the
chase). I'm tired of it and want to move on to other things (which I can
since starting flat seems to be reliable enough (if I've seen any
lockups, it was once or twice over two years)), but leaving some notes
seems like a good idea, especially regarding the apparent trigger. I'm
pretty certain (though not entirely convinced) it's a sneaky (VVL misses
it) synchronization error rather than a driver or hardware issue, though
things did seem to work nicely on AMD when Charles tried it for me (but
it was getting late and we were running into other issues with QF).