[qfcc] Add phantom uses to ud chains
Phantom uses are when when a partial write is done to a larger variable
(eg, a component in a vector, element in an array) after a write to the
whole variable, thus effectively using the definitions of the other
members of the variable. There may be a problem with nested structs, but
I'll worry about that when a test case comes up.

This fixes the alias_vec_live test. All tests pass again :)