[qfcc] Use correct type for return temps
If the compound expression specifies a type, or if the return value is a
geometric algebra type, that type should be used for creating the
initialized temp rather than the return type. This catches type errors
in GA (definitely) and will likely lead to fewer surprises for compound
expressions in general.