[qfcc] Recover gracefully from more errors
Avoids segfaults and internal errors when failing to compile a shader
when targeting the Ruamoko VM instead of SPIR-V (not that that works
anyway, but nobody likes segfaults or internal errors).

What's probably the non-obvious change is copying the parent scope's
space into the bypass scope. This isn't needed normally, but if an
undefined symbol error occurs while parsing a generic function, an int
def is created and it needs a place to go.