[qfcc] Mark processed compound initializers
The spir-v target doesn't need to create temps, so it needs only to
process a compound initializer into a state suitable for later emission.
However, this resulted in the initializer getting processed twice when
passed as an argument. Marking a compound initializer as already
processed (ie, it *is* an initialized temp) prevents the double
processing and thus the missing designated initializers error for
algebra types.