[util] Use dirent.c for MXE builds
That... was a mess. MXE provides readdir and firneds, but we were using
our own dirent.h, though the biggest problem was its lack of the d_type
member (very handy). However, fixing that resulted in weird crashes that
turned out to be the search handle having the wrong type (long... which
is only 32-bits on windows but the correct type is intptr_t). And then
there was the mess of TCHAR (utf-8 everywhere!). Well, it's pretty
ancient, so understandable.