[vulkan] Get brush gizmos mostly working
The shape is correct, but volumetrics aren't so it looks rather flat. GA
is making me relearn geometry, so basic things like point-plane
relations are still a little confusing, and line-plane relations
(especially getting the intersection point) even more so due to signs.
Also, it doesn't help I'd forgotten to edit the normal for the last
plane of the tetrahedron so it was essentially inside-out.