[model] Use a single alloc for skin textures
Multiple can fail when the staging buffer is nearly full and I don't
feel like trying to split up the commands. For some reason, even one
fails sometimes (when the gpu is slow to respond due to something else
using it) though the packet code is supposed to wait. I suspect there's
another bug in there somewhere since I did not expect a 32MB buffer to
be loaded to capacity with my test scene.