[qfcc] Do NOT use statement aux def if def not in operands
This essentially reverts 2be9821a595548fc001ae8d746fe983f45abb584, but I
don't want to revert it because the other changes in that commit might
still be relevant (or conflict). However, this does NOT break the printf
that the old commit fixed: it seems more recent improvements fixed it by
other means.

It turns out that using the statement aux def breaks (very badly) inout
(possibly out, too) parameters for void functions: the first (last?)
inout arg gets attached to the call, and then killed (I guess), and a
subsequent use of the inout arg uses the call directly, but it's void so
a null operand gets into the statement and then the compiler ICEs.

I don't know why (and this is a little concerning), but this also fixes
the inout_unused test (though it doesn't fix the *intended* test for
inout_void, which is probably a good thing: fixing 3 bugs with one
change would be a little too spooky).