[qfcc] Use inout/out params when for implicit return
It turns out the inout_void bug was due to the lack of an explicit
return statement at the end of the function: adding one (to test) fixed
it. Thus the real fix was simply doing the inout/out param handling for
the implicit return added by check_final_block.