[qfcc] Add a new failing test for ptrmove
It seems this is the bug that 2be9821a595548fc001ae8d746fe983f45abb584
was trying to fix (and did at the time). It's reversion in
63d44817e7d7f856cfc73a95b5128045e0dba7ae does not seem to be the cause
as putting the code back in did NOT fix the problem (and I don't want
that exact code anyway since removing it fixed several other issues). It
does look to be a problem with defines and/or kills via pointers.