[vulkan] Fix offset direction in expand shader
While the z_up transform is embedded in the view matrix, it wasn't
enough because of the screen aspect ratio, which is in the projection
matrix. I probably didn't need to do a full 4x4 matrix mult, but this
works. Also, I realized that since the offset is meant to be in pixels,
it has to take the perspective divide into account. This fixes some
nastily chopped up capsules I had seen.