[qfcc] Be less aggressive killing nodes
With the external memory checking for pointers, the need to kill
everything is much reduced. There may be some bugs with pointer writes
to local variables, but I think that's covered by the local checks. All
tests pass, at least, and qwaq seems to work. This is a part of trying
to improve array indexing, but I still get lots of reloads despite the
expressions no longer getting killed. I guess maybe something up with
finding nodes for lea.