[qfcc] Implement bitfield access for ruamoko
And probably v6p. Supporting bitfields for v6 is theoretically possible,
but the current implementation (at the language level) requires an int
rather than a float for the "backing" type, but insert and extract are
built using binary_expr() so it should be just a matter of tweaking the
type checks when declaring bitfields. Even sign is extracted.