[vulkan] Use bitfields for light data
So far, only on the *GPU* (I need to update the C code), but it served
as a nice test for reading bitfields. And yeah, that was worth the
effort despite the reason for actually doing bitfields was so I can
reuse qfmodel.h in Ruamoko (but I had wanted bitfields in shaders for a
while, just not enough to do anything about it).