[vulkan] Use explicit type for rd
It turns out using `auto` picks up the wrong reference type when
initializing from a buffer variable. I need to think about this a bit,
but I feel this is a bug in qfcc and it should do the right thing since
`auto` really works only for local variables in functions. It hasn't
been a problem before now because this is the first I've used it with a
struct in a buffer.