[qfcc] Fix pointer store analysis
I had goofed and copied the pointer move code forgetting I needed to
make it look like the pointer load code for for writes. This does not
fix false positives for uninitialized arrays when initialized by for
loops, but it does fix them when initialized in do-while loops (the
problem is the conditional entry to the for loop not being seen as
actually unconditional).