[qfcc] Split certain loop head nodes in two
Loop heads that conditionally enter the loop but the test is based on
constants or values initialized to a constant can be checked whether
they always enter the loop, or never, based on the test and initialized
values (eg, `for (int i = 0; i < 4; i++) {}`). Loop heads that are not
split are tight loops (head node jumps to itself), multi-loop heads (I
don't know if they exist, but I don't want to deal with them), loops
that are already entered unconditionally, and any where initial loop
entry can't be proven invariant. The test in the head is not yet
constant-folded: this commit only splits the head.