[ruamoko] Load a generated icosahedron
Annoyingly, having two qfm_model_t definitions with different struct
layouts results in an engine crash, so I had to change the definition in
armature (but that was, really, wrong anyway). There are some warts with
the mesh: it doesn't get unloaded properly and the default skin is
full-bright white. Also, it's smooth shaded, but it worked as a test.