[vulkan] Update dlight matdata correctly
While 6 matdata items per light need to be updated (one for each
matrix), it's usually not a good idea to instead update 6 times as many
lights as needed. Also, the update needs to be done after the light ids
for the dlights are set. It took overkill to trigger the problem.

I think there are still issues with some light updates as bigass1 has
some weird lighting in places: possibly some light vis issues, but I've
been pretty sure there are light vis issues for a while.