[qfcc] Check size when looking for phantom uses
If the alias is the same size as the main var, it's not a phantom use
since all possible members are written. While this fixes the
misidentification of an iterator var, I'm not sure it's *quite* the
right fix as the write was to the main var, not the alias. However, the
var is checked for being an alias? Still, it's not the wrong fix due to
the above reasoning.