[qwaq] Create a quadsphere mesh
For now, just the edges are shown using gizmos (ouch, expensive at high
subdivisions), but the quadsphere is made starting with a cube then
applying 5 levels of Catmull-Clark subdivision and normalizing the
vertex vectors. I used the half-edge algorithm found in
https://onrendering.com/data/papers/catmark/HalfedgeCatmullClark.pdf
Unfortunately, there's an error in the paper's calculation of vertex
count for a subdivision level (equation 7), but I was able to extract
the correct calculation by using equation 10 (which is correct) on two
different levels and subtracting. However, half-edges are nice :)